Finish rebase up to same functionality
The rebase of the HTML generation has reached the point of same
functionality as the previous version.

The attempt to create a DOM manager in python was removed. Instead there
are functions and classes to build/format the required HTML pages. Since
this is just a script to generate HTML pages, it should be easier to
maintain this version.