Fixed improper formatting with indent function
The indenter function would add indents at the end of the string block
(after the last new line), which would throw off the indents for any
text that came after. This was given a quick fix, by removing the
element created after the last newline, and adding a newline back at the
end once all the indentation is done.