Add functionality for creating Evolution section
To support having the Pokemon class build the HTML elements for the
evolution section, a small table will be built of just the pokemon
evolutions in the family line (from the full table). The small table is
then passed to the Pokemon class.
The Evolution section is built as a list of lists, as this is used to
create a family tree of sorts. Each stage in the evolutionary line is a
different level.
The way the lists are built is a precursor to a concept of a Family tree
class where each Pokemon is a node, and they are connected via an edge
that includes the evolution method. Currently it is just a connection of
lists.