Add generation of Transformation sections
Some Pokemon can change forms and the code was updated to show it in
their HTML page. The data is based on the "transformations" section in
the respective yaml file.
The function added handles the generation by itself. For simplicity in
creating the first iteration of the function, no code reuse/refactoring
was done.