Fix empty line in generated HTML files
There was a an empty line in the middle of the generated HTML files due
to the "Forms" section being empty. The indent helper function would add
a newline add the end even if the string was empty. Now the helper
function will return an empty string when the input is empty.