Update style/design of the Evolution tree
The evolution section was updated to add more detail to the tree. Each
Pokemon is given its own block, and indentation. The stylesheet was
updated to support this.
Since each evolution stage is underneath its previous stage, it adds
structure to the document for the tree (similar to nested lists).