Added methods to evolution tree on Pokemon page
The evolution section on a Pokemon's page now will have the method for
each evolution next to the Pokemon that is the next stage. To help
support this a helper function was made to format the dictionaries of
the method sections from the yaml file to a normal sentence.
Additionally the stylesheet was updated to include the arrow between the
method and stage.