Replace stat table HTML layout with grid/svg
The method for building the stats table was changed from using an HTML
table elements to using a grid display with css. The idea was that it
would give more potential control, and not be confined to the aspects of
the table element. There is not a dramatic difference in rendering.
Additionally drawing the bar to represent the stat was changed to be a
svg (rect) instead of using the css calc() function.