Fixed issue with base stage split evolutions
There was an issue with the evolution tree display for Pokemon that were
a third stage where there was a split at the base stage (ie Nidoran
line). This was due to a minor typo where when creating the table the
evolution stage was added to be checked next instead of the
pre-evolution, leading it to where the pre-evolution wasn't checked and
the split wasn't created.