Generates evolution tree order based on index
The display of the evolution tree is now based on the index of each
Pokemon in the tree/level. This only affected split evolutions, as
single path evolutions were always in the correct order. Previously
split evolution lines could have a different order of display between
pages.
The indexes of each Pokemon in the tree are gathered and passed on as
the end of the family object. Then it is popped off at the start of
generating the tree to not interfere with the previously established
code.