Consolidated code for creating family sub table
The code to create the sub table of evolutions for Pokemon family was
consolidated. Instead of breaking up the code into current Pokemon, and
then doing the same code but using a data structure, it adds the current
Pokemon to the data structure, removing the initial code.
Additional comments were made for clarity.