Moved evolution tree generation to main
The evolution tree generation was moved out of the Pokemon class and
into the main generation. This is a step towards consolidation with the
evolution related data structure generation.
Now the Pokemon class only has to generate HTML.
Some names of variables were updated for clarity in the new context.