Updated data structure for evolution tree
The data structure was changed from a list of lists to utilizing
dictionaries to represent the Pokemon. Each dictionary/Pokemon has its
name, filename, index and a list that links to the evolutions and their
methods. This provides the information directly to the Pokemon when
building the HTML, making the code simplier.
Additionally this means the family/evolution dictionary is similar to a
tree, where the evos list is the edges, and each Pokemon in the
dictionary is a node.